requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appnope==0.1.2
    # via ipython
backcall==0.2.0
    # via ipython
brotlipy==0.7.0
    # via -r requirements.in
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via brotlipy
chardet==4.0.0
    # via requests
cycler==0.10.0
    # via matplotlib
decorator==4.4.2
    # via ipython
dj-database-url==0.5.0
    # via -r requirements.in
dnspython==2.1.0
    # via -r requirements.in
fuzzywuzzy[speedup]==0.18.0
    # via -r requirements.in
geographiclib==1.50
    # via geopy
geopy==2.1.0
    # via -r requirements.in
idna==2.10
    # via requests
ipython-genutils==0.2.0
    # via traitlets
ipython==7.21.0
    # via -r requirements.in
jedi==0.18.0
    # via ipython
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.3.4
    # via -r requirements.in
numpy==1.20.1
    # via
    #   matplotlib
    #   pandas
    #   scipy
pandas==1.2.2
    # via -r requirements.in
parso==0.8.1
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.1.1
    # via matplotlib
prompt-toolkit==3.0.16
    # via ipython
psutil==5.8.0
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pycparser==2.20
    # via cffi
pygments==2.8.0
    # via ipython
pymongo==3.11.3
    # via -r requirements.in
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via
    #   matplotlib
    #   pandas
python-decouple==3.4
    # via -r requirements.in
python-levenshtein==0.12.2
    # via fuzzywuzzy
pytz==2021.1
    # via pandas
requests==2.25.1
    # via -r requirements.in
scipy==1.6.1
    # via -r requirements.in
six==1.15.0
    # via
    #   cycler
    #   python-dateutil
traitlets==5.0.5
    # via ipython
urllib3==1.26.3
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.36.2
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
